% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap_and_scale_fit}
\alias{cap_and_scale_fit}
\title{Cap and log-scale a response vector}
\usage{
cap_and_scale_fit(y, q = 0.995, eps = 1e-08)
}
\arguments{
\item{y}{numeric response (e.g., ADT counts)}

\item{q}{upper quantile for capping (default 0.995)}

\item{eps}{Small positive constant used for numerical stability.}
}
\value{
scaled numeric vector
}
\description{
Cap and log-scale a response vector
}
