% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMLE.R
\name{fmle_predict_mt}
\alias{fmle_predict_mt}
\title{Predict from a multi-task FMLE model}
\usage{
fmle_predict_mt(model, X_new, Z_new = NULL, return_se = TRUE)
}
\arguments{
\item{model}{object of class \code{"fmle_mt"}}

\item{X_new}{new feature matrix}

\item{Z_new}{new gating feature matrix}

\item{return_se}{logical}
}
\value{
list with \code{mean} (n x T), \code{alpha}, and optional \code{var}, \code{se}
}
\description{
Predict from a multi-task FMLE model
}
