% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMLE.R
\name{fmle_train_mt}
\alias{fmle_train_mt}
\title{Train FMLE model for a multi-task}
\usage{
fmle_train_mt(
  X,
  Y,
  Z = NULL,
  R = 6,
  m = 1.8,
  lambda_l1 = 0,
  ridge = 1e-06,
  standardize = TRUE,
  fcm_max_iter = 200,
  fcm_tol = 1e-05,
  seed = 1,
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix, n x p features}

\item{Y}{matrix, n x T responses (T tasks, e.g., multiple proteins)}

\item{Z}{matrix, n x d gating features}

\item{R}{number of experts}

\item{m}{fuzzifier}

\item{lambda_l1}{penalty (scalar, length-R, length-T, RxT matrix, or list)}

\item{ridge}{ridge penalty}

\item{standardize}{logical}

\item{fcm_max_iter}{max FCM iterations}

\item{fcm_tol}{tolerance}

\item{seed}{random seed}

\item{verbose}{logical}
}
\value{
object of class \code{"fmle_mt"}
}
\description{
Train FMLE model for a multi-task
}
